Spacing update
Overhauled the spacing on almost all sectional elements so that there will be space between the content and the edge of the screen on smaller browsers. Several edits was made to the margin and padding of the header area, so I might seperate this code into a new file in case it needs to be disabled.