Add new banner image CSS
We're going to eventually move away from loading the banner image from articles with Javascript and use modules instead. This marks code to be removed in the future and adds the new code going forward.

We'll remove the depreciated code when all banner image modules have been created and main article images removed